requirements_test.txt
mock==1.0.1
pytest==3.2.1
pytest-cov==2.5.1
hypothesis==3.27.1
pycodestyle==2.2.0
coverage==4.3.4
pynt==0.8.1
pyyaml>=4.2b1
coveralls==1.5.1
awscli>=1.16.85
cfn-lint>=0.11.1